+++ cuda.cuh
@@ -6,9 +6,9 @@ extern "C" {
 #endif
 
 void cuda_matmul(
-    int row_a, int col_a, double* data_a,
-    int row_b, int col_b, double* data_b,
-    int row_c, int col_c, double* data_c
+    int row_a, int col_a, float* data_a,
+    int row_b, int col_b, float* data_b,
+    int row_c, int col_c, float* data_c
 );
 
 #ifdef __cplusplus
