+++ main.c
@@ -20,11 +20,11 @@ int main(void) {
 
     int N_test = 1000;
     Matrix* X_test = matrix_cols(dataset->X, 0, N_test);
-    double* Y_test = dataset->Y;
+    float* Y_test = dataset->Y;
 
     int N_train = dataset->N - N_test;
     Matrix* X_train = matrix_cols(dataset->X, N_test, dataset->N);
-    double* Y_train = &dataset->Y[N_test];
+    float* Y_train = &dataset->Y[N_test];
 
     Parameters* params = gradient_descent(X_train, Y_train, N_train, LEARNING_RATE, EPOCHS);
     eval(X_test, Y_test, N_test, params);
